+++ sigle/src/modules/editor/TipTapEditor.tsx
@@ -23,6 +23,7 @@ import {
   SlashCommandsCommand,
 } from './extensions/SlashCommands';
 import { BubbleMenu } from './BubbleMenu';
+import { slashCommands, SlashCommandsList } from './InlineMenu';
 import { FloatingMenu } from './FloatingMenu';
 import {
   Heading1Light,
@@ -35,8 +36,6 @@ import { generateRandomId } from '../../utils';
 import { resizeImage } from '../../utils/image';
 import { storage } from '../../utils/blockstack';
 import { Story } from '../../types';
-import { colors } from '../../utils/colors';
-import { Flex, Text } from '../../ui';
 import { styled, globalCss, keyframes } from '../../stitches.config';
 
 const fadeInAnimation = keyframes({
@@ -67,44 +66,7 @@ const StyledEditorContent = styled(EditorContent, {
   },
 });
 
-const CommandsListItem = styled('li', {
-  cursor: 'pointer',
-  display: 'flex',
-  alignItems: 'center',
-  width: '100%',
-  py: '$1',
-  px: '$3',
-  '&.is-selected,&:hover': {
-    backgroundColor: '$gray3',
-  },
-});
-
-const CommandsList = (props: {
-  items: SlashCommandsCommand[];
-  selectedIndex: number;
-  selectItem: (index: number) => void;
-}) => {
-  const { items, selectedIndex, selectItem } = props;
-
-  return (
-    <Flex gap="2" direction="column">
-      {items.map(({ title, description, icon: Icon }, idx) => (
-        <CommandsListItem
-          key={idx}
-          className={selectedIndex === idx ? 'is-selected' : ''}
-          onClick={() => selectItem(idx)}
-        >
-          <Icon width={35} height={35} />
-          <Flex direction="column" css={{ ml: '$2' }}>
-            <Text>{title}</Text>
-            <Text css={{ color: '$gray9' }}>{description}</Text>
-          </Flex>
-        </CommandsListItem>
-      ))}
-    </Flex>
-  );
-};
-
+// Tippyjs theme used by the slash command menu
 const globalStylesCustomEditor = globalCss({
   ".tippy-box[data-theme~='sigle-editor'] .tippy-content": {
     padding: 0,
@@ -170,71 +132,8 @@ export const TipTapEditor = ({}: TipTapEditorProps) => {
       }),
       // Custom extensions
       SlashCommands.configure({
-        commands: [
-          {
-            icon: Heading1Light,
-            title: 'Heading 1',
-            description: 'Large section Heading',
-            command: ({ editor, range }) => {
-              editor
-                .chain()
-                .focus()
-                .deleteRange(range)
-                .setNode('heading', { level: 1 })
-                .run();
-            },
-          },
-          {
-            icon: Heading2Light,
-            title: 'Heading 2',
-            description: 'Medium section Heading',
-            command: ({ editor, range }) => {
-              editor
-                .chain()
-                .focus()
-                .deleteRange(range)
-                .setNode('heading', { level: 2 })
-                .run();
-            },
-          },
-          {
-            icon: Heading3Light,
-            title: 'Heading 3',
-            description: 'Small section Heading',
-            command: ({ editor, range }) => {
-              editor
-                .chain()
-                .focus()
-                .deleteRange(range)
-                .setNode('heading', { level: 3 })
-                .run();
-            },
-          },
-          {
-            icon: ImageLight,
-            title: 'Image',
-            description: 'Upload from your computer',
-            command: ({ editor, range }) => {
-              // TODO get it working
-              fileUploaderRef.current?.click();
-            },
-          },
-          {
-            icon: QuoteLight,
-            title: 'Quote',
-            description: 'Display a quote',
-            command: ({ editor, range }) => {
-              editor
-                .chain()
-                .focus()
-                // Use deleteRange to clear the text from command chars "/q" etc..
-                .deleteRange(range)
-                .toggleBlockquote()
-                .run();
-            },
-          },
-        ],
-        component: CommandsList,
+        commands: slashCommands,
+        component: SlashCommandsList,
       }),
     ],
     content: '<p>Hello World! 🌎️</p>',
